+++ ipmiTools/inc/ipmiHardware.h
@@ -3,7 +3,7 @@
  * @Date         : 2025-02-06 17:16:56
  * @Encoding     : UTF-8
  * @LastEditors  : stoneBeast
- * @LastEditTime : 2025-03-12 14:26:59
+ * @LastEditTime : 2025-04-09 15:03:26
  * @Description  : 声明ipmi相关硬件接口
  */
 
@@ -23,9 +23,9 @@ uint8_t __USER_IMPLEMENTATION read_flash(uint16_t addr, uint8_t read_len, uint8_
 uint8_t __USER_IMPLEMENTATION write_flash(uint16_t addr, uint8_t write_len, uint8_t* data);
 uint32_t __USER_IMPLEMENTATION get_sys_ticks(void);
 
-uint8_t __USER_IMPLEMENTATION read_sdr1_sensor_data(void);
-uint8_t __USER_IMPLEMENTATION read_sdr2_sensor_data(void);
-uint8_t __USER_IMPLEMENTATION read_sdr3_sensor_data(void);
-uint8_t __USER_IMPLEMENTATION read_sdr4_sensor_data(void);
+uint16_t __USER_IMPLEMENTATION read_sdr1_sensor_data(void);
+uint16_t __USER_IMPLEMENTATION read_sdr2_sensor_data(void);
+uint16_t __USER_IMPLEMENTATION read_sdr3_sensor_data(void);
+uint16_t __USER_IMPLEMENTATION read_sdr4_sensor_data(void);
 
 #endif // !__IPMI_HARDWARE_H
